==> notebooks/1_preprocessing.ipynb <==
# Предобработка датасета Adult

import pandas as pd

adult_data = pd.read_csv('../data/adult.csv')
adult_data.head()

## Пропуски данных

Для начала узнаем столбцы, в которых присутствуют пропуски данных

adult_data.isnull().sum()

Все столбцы с пропусками являются категориальными. При удалении строк с пропусками в худшем случае у нас останется 87% строк от исходного количества в датасете. В таком случае просто удаляем строки с пропусками и смотрим информацию о датасете после изменений

adult_data = adult_data.dropna()
adult_data.info()
adult_data.isnull().sum()

Все строки с пропуском данных убраны и у нас осталось более 92.6% от изначального объема датасета

## Порядковые признаки

Столбец education является порядковым признаком, а столбец education-num по сути является уже перекодированным c помощью ordinal encoding столбцом education, так что просто убираем столбец education

adult_data = adult_data.drop('education', 1)
adult_data.head()

## Категориальные признаки

К категориальными признакам относятся workclass, marital-status, occupation, relationship, race, native-country. Преобразуем их используя one-hot encoding

adult_data = pd.get_dummies(adult_data,prefix=['workclass'], columns=['workclass'])
adult_data = pd.get_dummies(adult_data,prefix=['marital-status'], columns=['marital-status'])
adult_data = pd.get_dummies(adult_data,prefix=['occupation'], columns=['occupation'])
adult_data = pd.get_dummies(adult_data,prefix=['relationship'], columns=['relationship'])
adult_data = pd.get_dummies(adult_data,prefix=['race'], columns=['race'])
adult_data = pd.get_dummies(adult_data,prefix=['native-country'], columns=['native-country'])

adult_data.head()

## Бинарные признаки

К бинарным признакам относятся: sex и salary. Воспользуемся ordinal encoding. Обозначим male как 0, female как 1. И для salary <=50K как 0 и >50K как 1

sex_mapping = {'Male': 0,'Female': 1}
salary_mapping = {'<=50K': 0, '>50K': 1}

adult_data['sex'] = adult_data['sex'].map(sex_mapping)
adult_data['salary'] = adult_data['salary'].map(salary_mapping)

adult_data.head()

## Cохранение в csv

adult_data.to_csv('../data/adult_preprocessed.csv', index=False)

# Предобработка датасета Machine

machine_data = pd.read_csv('../data/machine.csv')
machine_data.head()

## Пропуски данных

Для начала узнаем столбцы, в которых присутствуют пропуски данных

machine_data.isnull().sum()

Пропуски данных отсутсвуют, можно переходить к перекодировке

## Категориальные признаки

Категориальным признаком является vendor. Преобразуем его используюя one-hot encoding

machine_data = pd.get_dummies(machine_data,prefix=['vendor'], columns=['vendor'])
machine_data.head()

## Признак model

В связи с тем, что все строки имеют уникальное значение данного признака имеет смысл его удалить

machine_data = machine_data.drop('model', 1)
machine_data.head()

## Сохранение в csv

machine_data.to_csv('../data/machine_preprocessed.csv', index=False)
