==> notebooks/3_preprocessing_and_regression.ipynb <==
# Подготовка данных

## Импорт данных

import pandas as pd
dataset = pd.read_csv('../data/machine_preprocessed.csv')
dataset.head()

## Аномалии

Исходя из графиков визуализации данных, выполненных в файле 1_visualization можно предположить, что в данных присутствуют аномалии, однако данный датасет составлялся не на основе сбора данных людей, а на основе конкретных технических характеристик процессоров, в связи с этим нельзя отбросить какие-либо данные

## Разбиение данных

Для начала разобьем данные на множество описаний объектов и множество меток:

X = dataset.drop('erp', 1).values
y = dataset['erp'].values

Разобьем на обучающую и тестовую выборки:

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state = 42)

# Регрессия

Функция для вывода значений оценки

from sklearn.metrics import mean_squared_error, mean_absolute_error

def predict_estimation(estimator, y_test, y_pred):
    print("R^2 = ", estimator.score(X, y))
    print("MSE = ", mean_squared_error(y_test, y_pred))
    print("RMSE = ", mean_squared_error(y_test, y_pred, squared = False))
    print("MAE = ", mean_absolute_error(y_test, y_pred))

## Linear Regression

Документация:

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html

Обучение модели:

from sklearn.linear_model import LinearRegression

lr = LinearRegression()
lr.fit(X_train, y_train)

Оценим качество модели:

predict_estimation(lr, y_test, lr.predict(X_test))

Выведем полученные и предполагаемые значения:

print(lr.predict(X_test))
print(y_test)

Для данного метода отсутсвуют гиперпараметры

## Decision Tree Regressor

Документация:

https://scikit-learn.org/stable/modules/generated/sklearn.tree.DecisionTreeRegressor.html

Обучение модели:

from sklearn.tree import DecisionTreeRegressor

dtr = DecisionTreeRegressor()
dtr.fit(X_train, y_train)

Оценим качество модели:

predict_estimation(dtr, y_test, dtr.predict(X_test))

Выведем полученные и предполагаемые значения:

print(dtr.predict(X_test))
print(y_test)

Найдем наилучие гиперпараметры:

from sklearn.model_selection import GridSearchCV
param_grid = {"criterion": ["mse", "mae"],
              "min_samples_split": [10, 20, 40],
              "max_depth": [2, 6, 8],
              "min_samples_leaf": [20, 40, 100],
              "max_leaf_nodes": [5, 20, 100],
              }
grid_cv_dtr = GridSearchCV(dtr, param_grid, cv=5).fit(X,y)
print("Best Hyperparameters:\n{}".format(grid_cv_dtr.best_params_))

## Lasso

Документация:

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Lasso.html#sklearn.linear_model.Lasso

Обучение модели:

from sklearn.linear_model import Lasso

lo = Lasso()
lo.fit(X_train, y_train)

Оценим качество модели:

predict_estimation(lo, y_test, lo.predict(X_test))

Выведем полученные и предполагаемые значения:

print(lo.predict(X_test))
print(y_test)

Найдем наилучие гиперпараметры:

param_grid = {'alpha':[0.02, 0.024, 0.025, 0.026, 0.03]}
grid_cv_lo = GridSearchCV(lo, param_grid).fit(X,y)
print("Best Hyperparameters:\n{}".format(grid_cv_lo.best_params_))

## Ridge

Документация:

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html#sklearn.linear_model.Ridge

Обучение модели:

from sklearn.linear_model import Ridge

rg = Ridge()
rg.fit(X_train, y_train)

Оценим качество модели:

predict_estimation(rg, y_test, rg.predict(X_test))

Выведем полученные и предполагаемые значения:

print(rg.predict(X_test))
print(y_test)

Найдем наилучие гиперпараметры:

param_grid = {'alpha':[200, 230, 250, 265, 270, 275, 290, 300, 500]}
grid_cv_rg = GridSearchCV(rg, param_grid).fit(X,y)
print("Best Hyperparameters:\n{}".format(grid_cv_rg.best_params_))

## ElasticNet

Документация:

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html#sklearn.linear_model.ElasticNet

Обучение модели:

from sklearn.linear_model import ElasticNet

en = ElasticNet()
en.fit(X_train, y_train)

Оценим качество модели:

predict_estimation(en, y_test, en.predict(X_test))

Выведем полученные и предполагаемые значения:

print(en.predict(X_test))
print(y_test)

Найдем наилучие гиперпараметры:

param_grid = {'alpha': np.arange(0.1,2,0.05), 'l1_ratio': np.arange(0.1, 1.1, 0.1)}
grid_cv_en = GridSearchCV(en, param_grid).fit(X,y)
print("Best Hyperparameters:\n{}".format(grid_cv_en.best_params_))

## Вывод

Максимальный коэффициент детерминации имеет Ridge, следовательно является лучшим для данной модели
