==> notebooks/5_neural_network.ipynb <==
# Импорт

import pandas as pd
from sklearn import datasets
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from modules import mlpclsf
import numpy as np
import warnings
warnings.filterwarnings('ignore')

## Датасет Титаник

titanic = pd.read_csv('../data/titanic.csv', sep = ',')
titanic.head()

titanic.drop(['PassengerId', 'Name', 'Ticket', 'Cabin'], axis=1, inplace=True)
titanic.fillna(titanic.mean(), inplace=True)
titanic.head()

titanic = pd.get_dummies(data=titanic, prefix='c', columns=['Sex', 'Embarked'])
titanic.head()

Разделим на обучающую и тестовую выборки:

titanic_X = titanic.drop('Survived', 1).values
titanic_y = titanic['Survived'].values
titanic_X_train, titanic_X_test, titanic_y_train, titanic_y_test = train_test_split(titanic_X, titanic_y, test_size=0.2, stratify=titanic_y)

Инициализируем MLP:

mlp = mlpclsf.MLP('Sigmoid', 'SGD', layers=[titanic_X.shape[1], 100, 100, len(np.unique(titanic_y))])

Обучение модели:

mlp.fit(titanic_X_train, titanic_y_train)

Проверим точность полученных результатов:

print(sum(np.argmax(mlp.predict(titanic_X_test), axis=1) == titanic_y_test) / titanic_y_test.shape[0])

## Датасет Ирис

iris = datasets.load_iris()

Разделим на обучающую и тестовую выборки:

iris_X = iris.data
iris_y = iris.target
iris_X_train, iris_X_test, iris_y_train, iris_y_test = train_test_split(iris_X, iris_y, test_size=0.2, stratify=iris_y)

iris_X

Нормализуем данные:

scaler = StandardScaler()
iris_X_train = scaler.fit_transform(iris_X_train)
iris_X_test = scaler.transform(iris_X_test)

Инициализируем MLP:

mlp = mlpclsf.MLP('Sigmoid', 'SGD', layers=[iris_X.shape[1], 100, 100, len(np.unique(iris_y))])

Обучение модели:

mlp.fit(iris_X_train, iris_y_train)

Проверим точность полученных результатов:

print(sum(np.argmax(mlp.predict(iris_X_test), axis=1) == iris_y_test) / iris_y_test.shape[0])

## Task 2. Датасет Adult

adult = pd.read_csv('../data/adult_preprocessed.csv')
adult.head()

adult_X = adult.drop('salary', 1).values
adult_y = adult['salary'].values

Разделим на обучающую и тестовую выборки:

adult_X_train, adult_X_test, adult_y_train, adult_y_test = train_test_split(adult_X, adult_y, test_size=0.20, stratify=adult_y)

Нормализуем данные:

scaler = StandardScaler()
scaler.fit(adult_X_train)
X_train = scaler.transform(adult_X_train)
X_test = scaler.transform(adult_X_test)

Инициализируем MLP:

mlp = mlpclsf.MLP('Sigmoid', 'SGD', layers=[adult_X.shape[1], 100, 100, len(np.unique(adult_y))])

Обучение модели:

mlp.fit(adult_X_train, adult_y_train)

Проверим точность полученных результатов:

print(sum(np.argmax(mlp.predict(adult_X_test), axis=1) == adult_y_test) / adult_y_test.shape[0])
